refactor: Consolidate tool routing logic from StructuredLLMHostAgent into base HostAgent
- Move _route_tool_calls_to_agents() helper method to base HostAgent
- Simplify StructuredLLMHostAgent to minimal differences from base class
- Remove duplicate code and consolidate shared functionality
- Keep StructuredLLMHostAgent as thin subclass that just changes sequencing

Part of simplification effort in issue #94
